fix: keep pkg/cli/run within package file-count limit
Merge run_builder_test.go into run_test.go so adding run_response_stream_renderer_test.go stays within the 15-file pkg-file-count cap.
